// App/src/app/models/items.ts
export interface Items {
   index: string;
   nazwa: string;
   jm: string;
   cena: number;
   cena_h: number;
   rabat: number;
   proc_vat: string;
   waznosc: number;
   tandem: string;
   promocja: boolean;
   stan_rano: number;
   stan: number;
   przychod_7dni: number;
   rozchod_7dni: number;
   wyswietlane: boolean;
   imageUrl: string;
}